Fixed the return paths, the problem was twofold
The first part of the problem was that the end of blocks were not pointing to the true end of the blocks.
To solve this I added a new feild to statement, that is called last index. Last index is the ast's len at the end of the parsing of the statement.
This resolves to the end of any expanded expressions, and works with the pratt. I atempted to use posiiton, but pratt was messy.

I do not like this solution, however it works and we are really behind.

The second problem was that not all cases were falling through when they should have.
To solve this I added a flag for falling through and made the return conditional.